Validate batch number count with Addenda
Closes #17
